gh-1469 Fix broken tests
