+++ stroom-util/build.gradle
@@ -27,10 +27,11 @@ dependencies {
     compile libs.ws_rs_api
 
     testCompile libs.assertj_core
-    testCompile libs.commons_lang
-    testCompile libs.junit_jupiter_api
-    testCompile libs.vavr
     testCompile libs.classgraph
+    testCompile libs.commons_lang
+    testCompile libs.guice_extension
+    testCompile libs.junit_jupiter_api
+    testCompile libs.dropwizard_validation
 
     // The following logging libs are needed when running junits outside dropwizard
     testRuntimeOnly libs.jaxb_runtime
